+++ src/main/java/lagatrix/server/exceptions/manager/process/ProcessException.java
@@ -1,0 +1,17 @@
+package lagatrix.server.exceptions.manager.process;
+
+import lagatrix.server.exceptions.manager.ManagerException;
+
+/**
+ * This exception repersents all errors related to the process.
+ *
+ * @author javier
+ * @since 1.0
+ */
+public class ProcessException extends ManagerException {
+
+    public ProcessException(String description) {
+        super(description);
+    }
+
+}